fix the problems of uncorrect dynamic sd tbs
The previous implementation of dynamic sd tbs did not work correctly.
Seg fault in device code seemed to occur. In small GPU where the count
of SM is not large, the problem was latent. However, volta where the
total count of SM is 80 always suffered from kernel halt. To fix this
problem, followings were improved.
- fix wrong mAT structure
- add volatile
- fix sm allocation of kernel scheduling